backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# HTTP & API clients
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Data & Finance
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.33
alpha-vantage==2.3.1
ccxt==4.1.56

# Blockchain
web3==6.11.3
eth-account==0.10.0
hexbytes==0.3.1

# Task Queue
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx-mock==0.3.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
